Some refactoring

+++ angular/src/app/dashboard/custom-dashboard/dashboard/dashboard.component.ts
@@ -1,18 +1,18 @@
 import {Component, Injector} from '@angular/core';
+import {FormBuilder, FormGroup, Validators} from '@angular/forms';
+import {ActivatedRoute, Router} from '@angular/router';
+import {CreateDashboardTileModalComponent} from '@app/edit/create-dashboard-tile-modal/create-dashboard-tile-modal.component';
+import {appModuleAnimation} from '@shared/animations/routerTransition';
+import {AppComponentBase} from '@shared/app-component-base';
 import {
   DashboardDto,
   DashboardServiceProxy,
   DashboardTileDto,
-  DashboardTileType,
   UpdateTilePositionDto,
 } from '@shared/service-proxies/service-proxies';
-import {FormBuilder, FormGroup, Validators} from '@angular/forms';
-import {ActivatedRoute, Router} from '@angular/router';
-import {appModuleAnimation} from '@shared/animations/routerTransition';
-import {BsModalRef, BsModalService} from 'ngx-bootstrap/modal';
-import {CreateDashboardTileModalComponent} from '@app/edit/create-dashboard-tile-modal/create-dashboard-tile-modal.component';
 import {NgGridStackOptions, NgGridStackWidget} from 'gridstack/dist/angular';
-import {AppComponentBase} from '@shared/app-component-base';
+import {BsModalRef, BsModalService} from 'ngx-bootstrap/modal';
+import {CONSTRAINTSBYTYPE, GRIDOPTIONS} from './dashboard.constants';
 
 @Component({
   selector: 'app-dashboard',
@@ -21,23 +21,18 @@ import {AppComponentBase} from '@shared/app-component-base';
   animations: [appModuleAnimation()],
 })
 export class DashboardComponent extends AppComponentBase {
-  editMode = false;
-  editName = false;
-  dashboard: DashboardDto;
-  editFormGroup: FormGroup;
   id: number;
-  createDashboardTileModal: BsModalRef;
+  dashboard: DashboardDto;
+
   gridOptions: NgGridStackOptions;
   gridWidgets: CustomGridStackWidget[];
+  constraintsByType = CONSTRAINTSBYTYPE;
 
+  editMode = false;
+  editName = false;
+  editFormGroup: FormGroup;
   isTrashVisible = false;
-
-  constraintsByTileType = {
-    [DashboardTileType._0]: {minW: 2, minH: 2, maxW: 6, maxH: 3},
-    [DashboardTileType._1]: {minW: 2, minH: 2, maxW: 6, maxH: 8},
-    [DashboardTileType._2]: {minW: 2, minH: 2, maxW: 6, maxH: 4},
-    [DashboardTileType._3]: {minW: 1, minH: 2, maxW: 12, maxH: 8},
-  };
+  createDashboardTileModal: BsModalRef;
 
   constructor(
     injector: Injector,
@@ -57,41 +52,33 @@ export class DashboardComponent extends AppComponentBase {
       this.loadDashboard();
     });
 
-    this.gridOptions = {
-      margin: 5,
-      float: true,
-      removable: '.trash',
-      cellHeight: '3.3rem',
-      columnOpts: {
-        columnWidth: 100,
-        columnMax: 12,
-        layout: 'scale',
-      },
-      minRow: 19,
-    };
+    this.gridOptions = GRIDOPTIONS;
   }
 
   loadDashboard() {
     this._dashboardService.get(this.id).subscribe(result => {
       this.dashboard = result;
       this.editFormGroup.controls.name.setValue(this.dashboard.name);
-      this.initializeGrid();
+
+      this.gridWidgets = this.dashboard.tiles.map(tile =>
+        this.createWidget(tile)
+      );
     });
   }
 
-  initializeGrid() {
-    this.gridWidgets = this.dashboard.tiles.map(tile => ({
+  createWidget(tile: DashboardTileDto): CustomGridStackWidget {
+    return {
       x: tile.x,
       y: tile.y,
       w: tile.width,
       h: tile.height,
-      minW: this.constraintsByTileType[tile.itemType].minW,
-      minH: this.constraintsByTileType[tile.itemType].minH,
-      maxW: this.constraintsByTileType[tile.itemType].maxW,
-      maxH: this.constraintsByTileType[tile.itemType].maxH,
+      minW: this.constraintsByType[tile.itemType].minW,
+      minH: this.constraintsByType[tile.itemType].minH,
+      maxW: this.constraintsByType[tile.itemType].maxW,
+      maxH: this.constraintsByType[tile.itemType].maxH,
       id: tile.id.toString(),
       tile: tile,
-    }));
+    };
   }
 
   editNameClicked() {
@@ -150,47 +137,39 @@ export class DashboardComponent extends AppComponentBase {
       return;
     }
 
-    const positionDtos: UpdateTilePositionDto[] = [];
-    for (const node of $event.nodes) {
-      if (!this.hasTileMoved(node.tile, node.x, node.y, node.w, node.h)) {
-        continue;
-      }
-      const positionDto = new UpdateTilePositionDto();
-      positionDto.tileId = node.tile.id;
-      positionDto.x = node.x;
-      positionDto.y = node.y;
-      positionDto.width = node.w;
-      positionDto.height = node.h;
-
-      positionDtos.push(positionDto);
-    }
+    const positionDtos = $event.nodes.map(
+      node =>
+        new UpdateTilePositionDto({
+          tileId: node.tile.id,
+          x: node.x,
+          y: node.y,
+          width: node.w,
+          height: node.h,
+        })
+    );
 
     this._dashboardService
       .updateTilePositions(this.id, positionDtos)
       .subscribe(() => {
-        for (const node of $event.nodes) {
+        $event.nodes.forEach(node => {
           const tile = this.dashboard.tiles.find(t => t.id === node.tile.id);
-          tile.x = node.x;
-          tile.y = node.y;
-          tile.width = node.w;
-          tile.height = node.h;
-        }
+          if (tile) {
+            Object.assign(tile, {
+              x: node.x,
+              y: node.y,
+              width: node.w,
+              height: node.h,
+            });
+          }
+        });
       });
   }
 
-  private hasTileMoved(
-    tile: DashboardTileDto,
-    x: number,
-    y: number,
-    w: number,
-    h: number
-  ): boolean {
-    return (
-      tile.x !== x || tile.y !== y || tile.width !== w || tile.height !== h
-    );
-  }
-
   public onGridRemove($event: {event: Event; nodes: CustomGridStackWidget[]}) {
+    if (!$event.nodes || $event.nodes.length === 0) {
+      return;
+    }
+
     for (const node of $event.nodes) {
       this._dashboardService.deleteTile(this.id, node.tile.id).subscribe();
     }
